IR: Add support for break and continue statements
The front-end is already doing the work of connecting this statements to their "parent" statement, so we just needed to build a map from the `Stmt*` to the corresponding `IRBlock*`s to use for break/continue when outputting any loop statement, and then look up in the map for the branch target when outputting a break/continue.

When we get around to adding `switch` statements, the same pattern should work just fine.

I also added support for `do/while` statements in IR codegen, and made sure to exercise those in one of the test cases I added.

There is also an unrelated IR codegen fix for when there is a "bound subscript" on the RHS of an assignment.