IR: Add support for break and continue statements (#272)
* IR: Add support for break and continue statements

The front-end is already doing the work of connecting this statements to their "parent" statement, so we just needed to build a map from the `Stmt*` to the corresponding `IRBlock*`s to use for break/continue when outputting any loop statement, and then look up in the map for the branch target when outputting a break/continue.

When we get around to adding `switch` statements, the same pattern should work just fine.

I also added support for `do/while` statements in IR codegen, and made sure to exercise those in one of the test cases I added.

There is also an unrelated IR codegen fix for when there is a "bound subscript" on the RHS of an assignment.

* IR: fix handling of do/while and continue

Thanks to @csyonghe for pointing out my mistake in the earlier commit.
I implemented `continue` for `do/while` loops incorrectly, branching to the head of the loop instead of the loop test.
I'll try to blame this mistake on the fact that I never use `do/while` loops because I think they are awful. :)

The fix for that issue wasn't too bad (see `lower-to-ir.cpp`) but it surfaces a much more serious issue: I wasn't actually implementing `continue` correctly *at all* when it comes to generating HLSL/GLSL from the IR (I can't easily make an excuse for that one).
The basic issue at the heart of this is that given an input statement like:

```
for(int ii = 0; ii < N; ii = doSomething(ii))
{ ... }
```

The continue clause (`ii = doSomething(ii)`) could expand into many instructions (across multiple blocks, if we inline), and there is in general no guarantee when we are done that we can package up that code as an expression and spit out a new `for` loop (the same basic argument applies to a `do { ... } while(someComplexExpression())`.

So, if we assume that in general we have to generate a full *statement* for the `continue` clause, what can we emit?

- We could try to "outline" the continue code into its own function, so that we can call it from an expression. That could work, but has high implementation complexity.

- We could introduce additional `bool` variables for control flow, outputting something like:
  ```
  bool useContinueBlock = false;
  for(;;) {
    if(useContinueBlock) { <CONTINUE CODE>; }
    useContinueBlock = true;
    <LOOP TEST>
    <LOOP BODY>
  }
  ```
  This works but user might balk at the extra variable we introduce.

- We could duplicate the code at each continue site. That is, we emit the loop as:
  ```
  for(;;) {
    <LOOP TEST>
    <LOOP BODY>
    <CONTINUE CODE>
  }
  ```
  but then whenever we'd like to emit `continue;` we instead emit `{ <CONTINUE CODE>; continue; }`.
  This doesn't introduce any extra variables, but it causes code duplication (limited, if we don't have too many `continue` sites, and the continue clause is small - which are the common cases).

When I was initially working on the IR codegen I picked that last option just because it is what `fxc` seems to do, but I neglected to actually *implement* the special-case codegen for a `continue` instruction. This change addresses that (see `emit.cpp`).

Finally, once things were fixed the `continue` test case produced the results Yong told me to expect, but it also produced a warning from the downstream HLSL compiler ("hey, your loop doesn't ever actually *loop*!"), so I reworked the test back to one that actually loops (but still tests `continue`).

As a final aside in this essay of a commit message: the current IR representation of control flow uses special-case instructions for various cases of unconditional branch (and two variations on `if`), but these are not strictly necessary, and a future change will hopefully clean it up. The biggest catch in doing that is that it will require the IR->source codegen to carefully track which blocks represent which kinds of branch targets in context (e.g., you can't assume that a `continue` that nees the special handling above will appear as a distinct kind of instruction).