Add sample-rate-input detection for HLSL.
In the HLSL case, it is possible to do this detection entirely based on declared signatures (it doesn't have a dependency on code generation like the GLSL case does). I've added test cases for the two main ways that a shader can become sample rate:

1. Qualify a fragment input with `sample`
2. Accept an input with the `SV_SampleIndex` semantic

In each case I nested the input inside a `struct` to try to match common HLSL idiom, and to make sure that we handle the nested case.

This code is *not* robust against shaders that declare such an input and then never use it, but that is to be expected given the goals for Slang.