Fix up parameter block emit for mixed rewriter+IR (#316)
* Fix up parameter block emit for mixed rewriter+IR

The basic problem here arose when a parameter block was declared in code that will be lowered via IR, but is referenced from user code that will use the AST-based lowering pass. The type legalization would split up the parameter block, and the AST would refer to the new variables, but it would fail to recognize when those variables represent constant buffers, and thus should get implicit-dereference behavior.

The fix here was just to propagate type information through when creating new AST (pseudo-)expressions to represent IR declarations that were split.

A more complete fix down the road should try to make sure that both the expression emit logic and the declaration-emit logic agree on whether a particular declaration of a parameter block or constant buffer needs to support the "implicit dereference" case or not. I'm leaving that to future work.

With this change, two test cases that had been disabled now pass.

* Fixup: don't do implicit deref for `ParameterBlock` values

Right now the rules for `ParameterBlock` and all other uniform praameter groups needs to be different, but the `isImplicitBaseExpr` logic wasn't taking that into account.