Support generic type constraints when implicitly invoking generic
Fixes #326

This basically just copy-pastes logic from the explicit case over to the implicit case.
After we've solved for the explicit type/value arguments, we loop over the constraints and for each one we try to find a suitable subtype witness to use (after substituting in the arguments solved so far).

This change includes a test case for the new functionality.